Criando endpoint visualizar cliente

+++ src/main/java/com/fiap/gerenciamento_clientes/infraestructure/ClienteJpaRepository.java
@@ -3,5 +3,5 @@ package com.fiap.gerenciamento_clientes.infraestructure;
 import com.fiap.gerenciamento_clientes.infraestructure.entities.ClienteEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClienteJpaRepository extends JpaRepository<ClienteEntity, Integer> {
+public interface ClienteJpaRepository extends JpaRepository<ClienteEntity, Long> {
 }
